+++ src/app/services/CreateAppointmentService.js
@@ -1,0 +1,79 @@
+/**
+ * posso usar esse service em varios lugares da app
+ * normalmente um service faz uma unica coisa, e pode ser utilizado em varios lugares
+ */
+import { startOfHour, parseISO, isBefore, format } from 'date-fns';
+import pt from 'date-fns/locale/pt-BR';
+
+import User from '../models/User';
+import Appointment from '../models/Appointment';
+
+import Notification from '../schemas/Notifications';
+
+class CreateAppointmentService {
+  async run({ provider_id, user_id, date }) {
+    if (provider_id === user_id) {
+      throw new Error("can't schedule for yourself");
+    }
+
+    /**
+     * Check if provider_id is a provider
+     */
+    const isProvider = await User.findOne({
+      where: { id: provider_id, provider: true },
+    });
+
+    if (!isProvider) {
+      throw new Error('You can only create appointments with providers');
+    }
+
+    /**
+     * Check for past dates
+     */
+    const hourStart = startOfHour(parseISO(date));
+    if (isBefore(hourStart, new Date())) {
+      throw new Error('Past dates are not permitted');
+    }
+
+    /**
+     * Check date availability
+     */
+    const checkAvailability = await Appointment.findOne({
+      where: {
+        provider_id,
+        canceled_at: null,
+        date: hourStart,
+      },
+    });
+
+    if (checkAvailability) {
+      throw new Error('Appointment date is not available');
+    }
+
+    /**
+     * Notificar provider
+     */
+    const user = await User.findByPk(user_id);
+    const formattedDate = format(
+      hourStart,
+      "'dia ' dd 'de' MMMM', às' H:mm'h'",
+      { locale: pt }
+    );
+
+    await Notification.create({
+      content: `Novo agendamento de ${user.name} para ${formattedDate}`,
+      user: provider_id,
+    });
+
+    const appointment = await Appointment.create({
+      user_id,
+      provider_id,
+      date,
+    });
+
+    return appointment;
+  }
+}
+
+// singleton
+export default new CreateAppointmentService();
